Add localSubnetOnly parameter
If localSubnetOnly is set, then the backup is only done if the destination is a local disk, or on a server file share in a locally-connected subnet.
It checks for destinations like:
- UNC path \\server.domain\backupshare
- Drive mapped with "net use Z: \\server.domain\backupshare"
- Drive mapped with "subst Z: \\server.domain\backupshare"
It can detect all of these, extract the "server.domain", lookup the IPv4 address, then compare that IPv4 address to all of the locally-connected subnets (e.g. LAN cable, wireless, 3G dongle, VPN tunnel) and decide if the server is local or not.
Thus we can do laptop backups to a "home office" server only when the laptop is actually in its "home office". When it is at another office across the VPN intranet (or off the intranet completely) the backup task will report that the destination share is not in a local subnet.