Allow Delorean backup to work with hidden folders
If the input folder for the backup is a hidden folder (like we are hiding D:\Tmail from the user) then the output backup on "X:\Backup\Tmail - datetime" is also a hidden file.
When looking for previous backups in backup destination like X:\Backup, we need to force Get-ChildItem to return all files, and then to see if it is a directory, the attributes just need to contain (-match) "Directory". For a hidden directory, Atrributes returns "Hidden, Directory".