Add emailJobName option
This allows the user to specify a little bit of text that goes in the email subject, along with the hostname and username. This is handy when there are multiple different backups for different purposes on the same host. It makes it easy to distinguish the different backup emails by email subject. e.g.
-emailJobName 'to NAS'
The email subject becomes something like:
Backup of: OFFICE-DC-01 to NAS by: backup.account

Also added output of backupMappedPath to the email body summary. That helps when drive X:\ is actually mapped to some file share \\office-nas.company.com\backup - then the user can see in the email body where the backup destination drive letter was actually mapped to.