Add noads noea, fix timeTolerance on copy
Add the functionality to pass in the "noads" and "noea" and put them onto the "ln" command line. This was needed during some testing to find out why some files on a "keep 1 copy only" backup to NAS kept being copied every time.
Actually the problem was that timeTolerance was not being put on the command line for the first backup (--copy). But it is needed for the case when keeping only 1 backup, when we "mirror" using the files that are already there. Now a "mirror"-style job to NAS works better.
Note: we should probably look at using the actual --mirror option for this case, rather than --copy - it might do some better things.
Note 2: The noads and noea did not help the case I was testing, but it seems useful to have the functionality there so we can use it when we want to or have to.