Support doing subst as parameter
This allows the user to specify a subst drive letter on the command line or in an ini file. This script can then do the work to "subst" the drive letter, and delete the "subst" at the end. It saves the user having to do it externally in an outer script. It also means the use of a "subst" drive letter with UNC backupDestination can be controlled from the ini file, giving more flexibility.
Also log the resulting backupMappedString in more places so it is always clear where the backup is really going.