Windows expects subnet mask in 255.255.255.0 format
localSubnetMask was an integer, and we were passing a number if bits (CIDR) like "16", "24". But the code at line 778/799:
[IPAddress]$mask = $localSubnetMask
did not work - that set the $mask var to be an IPAddress object with an internal value equivalent to 0.0.0.12 - not quite what we want!
The "easy" way is to succumb to the 255.255.255.0 style of specifying the network mask. That style drops straight into the [IPAddress\ object and the bit patterns that come out can be "anded" with the real IP addresses that are also in [IPAddress] object format and it all just works without having to know exactly what bit order the underlying structures are stored in.
This now seems to work fine - I have tried doing backup from home to office across various subnet mask sizes and it is making the expected decisions about whether to do the backup or not.