Improve error output when testing possible destinations
Issue
https://github.com/International-Nepal-Fellowship/Windows-Tools/issues/37
Problems with possible destinations (e.g. destination share does not
exist, server name not found...) are now logged as WARNING only. They
are written to the log file so someone can always see the process that
happened. They are only written to the email body if there was no
successful destination actually found. If one of the destinations worked
OK then we do not want to pollute the email body of a successful backup
with WARNING messages that will confuse the recipient.