Single copy should be mirror
When using "--copy" files accumulate - old files that have been deleted
in the source never get deleted from the destination.
Using "--mirror", the files in destination end up being exactly the same
as those in the source - new files are copied to, changed files are
overwritten in, deleted files are reomved from the destination.